+++ spec/run.lua
@@ -73,6 +73,7 @@ local files = {
 	"spec/normalize_test.lua",
 	"spec/format_test.lua",
 	"spec/brand_test.lua",
+	"spec/features_test.lua",
 	"spec/api_test.lua",
 	"spec/notify_test.lua",
 	"spec/model_test.lua",